+++ demisto_sdk/__main__.py
@@ -2287,13 +2287,20 @@ def error_code(config, **kwargs):
 )
 @click.option('-ud', '--use-docker', is_flag=True, help="Use docker service to run the content graph")
 @click.option('-us', '--use-existing', is_flag=True, help="Use existing service", default=False)
+@click.option('-d', '--dependencies', is_flag=True, help="Whether dependencies should be included in the graph", default=False)
 @click.option('-o', '--output-file', type=click.Path(), help="dump file output", default=None)
 @click.option('-v', "--verbose", count=True, help="Verbosity level -v / -vv / .. / -vvv",
               type=click.IntRange(0, 3, clamp=True), default=2, show_default=True)
 @click.option('-q', "--quiet", is_flag=True, help="Quiet output, only output results in the end")
 @click.option("-lp", "--log-path", help="Path to store all levels of logs",
               type=click.Path(resolve_path=True))
-def create_content_graph(use_docker: bool = False, use_existing: bool = False, output_file: Path = None, **kwargs):
+def create_content_graph(
+    use_docker: bool = False,
+    use_existing: bool = False,
+    dependencies: bool = False,
+    output_file: Path = None,
+    **kwargs,
+):
     from demisto_sdk.commands.common.logger import logging_setup
     from demisto_sdk.commands.content_graph.content_graph_commands import \
         create_content_graph as create_content_graph_command
@@ -2305,7 +2312,7 @@ def create_content_graph(use_docker: bool = False, use_existing: bool = False, o
         output_file=Path(output_file) if output_file else None,
         use_docker=use_docker,
     ) as content_graph_interface:
-        create_content_graph_command(content_graph_interface)
+        create_content_graph_command(content_graph_interface, dependencies)
 
 
 @main.result_callback()