+++ demisto_sdk/commands/content_graph/parsers/report.py
@@ -15,6 +15,10 @@ class ReportParser(JSONContentItemParser, content_type=ContentType.REPORT):
 
         self.connect_to_dependencies()
 
+    @property
+    def marketplaces(self) -> List[MarketplaceVersions]:
+        return [MarketplaceVersions.XSOAR]
+
     def connect_to_dependencies(self) -> None:
         """Collects scripts used in the report as optional dependencies."""
         for layout in self.json_data.get("dashboard", {}).get("layout", []):