+++ demisto_sdk/commands/generate_docs/tests/generate_docs_test.py
@@ -676,10 +676,13 @@ class TestAppendOrReplaceCommandInDocs:
     command = 'dxl-send-event'
     old_doc = open(positive_test_data_file).read()
     new_docs = "\n<NEW DOCS>\n"
+    new_command = '\n### dxl-send-event-new-one\n***\nSends the specified event to the DXL fabric.\n##### Base Command\n`dxl-send-event-new-one`'
+    new_command += '\n##### Input\n| **Argument Name** | **Description** | **Required** |\n| --- | --- | --- |\n| topic | The topic for which to publish the'
+    new_command += ' message. | Required |\n| payload | The event payload. | Required |\n##### Context Output\nThere is no context output for this command.'
     positive_inputs = [
-        (old_doc, new_docs),
-        (old_doc + "\n## Known Limitation", new_docs + "\n## Known Limitation"),
-        (old_doc + "\n### new-command", new_docs + "\n### new-command"),
+        (old_doc, new_docs + new_command),
+        (old_doc + "\n## Known Limitation", new_docs + new_command + "\n## Known Limitation"),
+        (old_doc + "\n### new-command", new_docs + new_command + "\n### new-command"),
         ("no docs (empty)\n", "no docs (empty)\n" + new_docs),
         (f"Command in file, but cant replace. {command}", f"Command in file, but cant replace. {command}\n" + new_docs)
     ]