+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * Fixed an issue where **validate** failed to recognize integration tests that were missing from config.json
 
 ## Unreleased
+* Fixed an issue where **generate-docs** with `-c` argument updated sections of the incorrect commands.
+* Added IF113 error code to **ALLOWED_IGNORE_ERRORS**.
 * Added support for marketplace-specific tags.
 
 ## 1.7.7